+++ packages/asc-ui/src/components/Input/InputStyle.ts
@@ -1,5 +1,6 @@
 import styled, { css } from 'styled-components'
 import { focusStyleOutline, themeColor, themeSpacing } from '../../utils'
+import { INPUT_SIZE } from '../shared/constants'
 
 type StyleProps = {
   error?: boolean
@@ -14,6 +15,7 @@ export const inputCss = css<StyleProps>`
   line-height: 18px;
   padding: ${themeSpacing(2)};
   width: 100%;
+  height: ${INPUT_SIZE}px;
   ${focusStyleOutline(2, 0.5)}
   ${({ error }) =>
     !error &&